HEMS/requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
Django==1.9.2
django-filter==1.1.0
djangorestframework==3.7.3
idna==2.6
kombu==4.1.0
Markdown==2.6.9
MySQL-python==1.2.5
pytz==2017.3
requests==2.18.4
urllib3==1.22
vine==1.1.4
